Moved endpoints to a separate file
